Bridge: add node_bridge_get_bytearray
