+++ rust/bridge/node/src/lib.rs
@@ -8,7 +8,6 @@ use libsignal_bridge::*;
 use libsignal_protocol_rust::*;
 use neon::context::Context;
 use neon::prelude::*;
-use std::convert::TryFrom;
 
 fn return_boolean<'a>(
     cx: &mut FunctionContext<'a>,
@@ -20,49 +19,12 @@ fn return_boolean<'a>(
     }
 }
 
-fn return_binary_data<'a, T: AsRef<[u8]>>(
-    cx: &mut FunctionContext<'a>,
-    bytes: Result<T, SignalProtocolError>,
-) -> JsResult<'a, JsValue> {
-    match bytes {
-        Ok(bytes) => {
-            let bytes = bytes.as_ref();
-
-            let bytes_len = match u32::try_from(bytes.len()) {
-                Ok(l) => l,
-                Err(_) => {
-                    return cx.throw_error("Cannot return very large object to JS environment")
-                }
-            };
-            let mut buffer = cx.buffer(bytes_len)?;
-            cx.borrow_mut(&mut buffer, |raw_buffer| {
-                raw_buffer.as_mut_slice().copy_from_slice(&bytes);
-            });
-            Ok(buffer.upcast())
-        }
-        Err(e) => cx.throw_error(e.to_string()),
-    }
-}
-
-macro_rules! node_bridge_serialize {
-    ( $typ:ident::$fn:ident is $node_name:ident ) => {
-        #[allow(non_snake_case)]
-        fn $node_name(mut cx: FunctionContext) -> JsResult<JsValue> {
-            let obj = cx.argument::<DefaultJsBox<$typ>>(0)?;
-            let bytes = obj.$fn();
-            return_binary_data(&mut cx, Ok(bytes))
-        }
-    };
-}
-
 #[allow(non_snake_case)]
 fn PrivateKey_generate(mut cx: FunctionContext) -> JsResult<JsValue> {
     let mut rng = rand::rngs::OsRng;
     let keypair = KeyPair::generate(&mut rng);
     return_boxed_object(&mut cx, Ok(keypair.private_key))
 }
-
-node_bridge_serialize!(PrivateKey::serialize is PrivateKey_serialize);
 
 #[allow(non_snake_case)]
 fn PrivateKey_getPublicKey(mut cx: FunctionContext) -> JsResult<JsValue> {
@@ -95,8 +57,6 @@ fn PrivateKey_agree(mut cx: FunctionContext) -> JsResult<JsValue> {
     return_binary_data(&mut cx, shared_secret)
 }
 
-node_bridge_serialize!(PublicKey::serialize is PublicKey_serialize);
-
 #[allow(non_snake_case)]
 fn PublicKey_verify(mut cx: FunctionContext) -> JsResult<JsValue> {
     let key = cx.argument::<DefaultJsBox<PublicKey>>(0)?;
@@ -117,13 +77,13 @@ fn PublicKey_verify(mut cx: FunctionContext) -> JsResult<JsValue> {
 fn main(mut cx: ModuleContext) -> NeonResult<()> {
     cx.export_function("PrivateKey_generate", PrivateKey_generate)?;
     cx.export_function("PrivateKey_deserialize", node_PrivateKey_deserialize)?;
-    cx.export_function("PrivateKey_serialize", PrivateKey_serialize)?;
+    cx.export_function("PrivateKey_serialize", node_PrivateKey_serialize)?;
     cx.export_function("PrivateKey_sign", PrivateKey_sign)?;
     cx.export_function("PrivateKey_agree", PrivateKey_agree)?;
     cx.export_function("PrivateKey_getPublicKey", PrivateKey_getPublicKey)?;
 
     cx.export_function("PublicKey_verify", PublicKey_verify)?;
     cx.export_function("PublicKey_deserialize", node_PublicKey_deserialize)?;
-    cx.export_function("PublicKey_serialize", PublicKey_serialize)?;
+    cx.export_function("PublicKey_serialize", node_PublicKey_serialize)?;
     Ok(())
 }
