+++ rust/bridge/shared/src/node/mod.rs
@@ -5,6 +5,7 @@
 
 use libsignal_protocol_rust::*;
 use neon::context::Context;
+use std::convert::TryFrom;
 use std::ops::Deref;
 
 pub use neon::prelude::*;
@@ -29,6 +30,30 @@ pub fn return_boxed_object<'a, T: 'static + Send>(
 ) -> JsResult<'a, JsValue> {
     match value {
         Ok(v) => Ok(cx.boxed(DefaultFinalize(v)).upcast()),
+        Err(e) => cx.throw_error(e.to_string()),
+    }
+}
+
+pub fn return_binary_data<'a, T: AsRef<[u8]>>(
+    cx: &mut FunctionContext<'a>,
+    bytes: Result<T, SignalProtocolError>,
+) -> JsResult<'a, JsValue> {
+    match bytes {
+        Ok(bytes) => {
+            let bytes = bytes.as_ref();
+
+            let bytes_len = match u32::try_from(bytes.len()) {
+                Ok(l) => l,
+                Err(_) => {
+                    return cx.throw_error("Cannot return very large object to JS environment")
+                }
+            };
+            let mut buffer = cx.buffer(bytes_len)?;
+            cx.borrow_mut(&mut buffer, |raw_buffer| {
+                raw_buffer.as_mut_slice().copy_from_slice(&bytes);
+            });
+            Ok(buffer.upcast())
+        }
         Err(e) => cx.throw_error(e.to_string()),
     }
 }
@@ -60,5 +85,29 @@ macro_rules! node_bridge_deserialize {
     };
     ( $typ:ident::$fn:path ) => {
         node_bridge_deserialize!($typ::$fn as $typ);
+    };
+}
+
+macro_rules! node_bridge_get_bytearray {
+    ( $name:ident($typ:ty) as None => $body:expr ) => {};
+    ( $name:ident($typ:ty) as $node_name:ident => $body:expr ) => {
+        paste! {
+            #[allow(non_snake_case)]
+            pub fn [<node_ $node_name>](
+                mut cx: node::FunctionContext
+            ) -> node::JsResult<node::JsValue> {
+                expr_as_fn!(inner_get<'a>(
+                    obj: &'a $typ
+                ) -> Result<impl AsRef<[u8]> + 'a, SignalProtocolError> => $body);
+                let obj = cx.argument::<node::DefaultJsBox<$typ>>(0)?;
+                let bytes = inner_get(&obj);
+                node::return_binary_data(&mut cx, bytes)
+            }
+        }
+    };
+    ( $name:ident($typ:ty) => $body:expr ) => {
+        paste! {
+            node_bridge_get_bytearray!($name($typ) as [<$typ _ $name>] => $body);
+        }
     };
 }
